[ConstraintSystem] Remove now completely obsolete CSDiag
