+++ lib/Sema/CMakeLists.txt
@@ -3,7 +3,6 @@ add_swift_host_library(swiftSema STATIC
   BuilderTransform.cpp
   CSApply.cpp
   CSBindings.cpp
-  CSDiag.cpp
   CSGen.cpp
   CSRanking.cpp
   CSSimplify.cpp
