--- lib/Sema/CSDiag.cpp
@@ -1,687 +1,0 @@
-//===--- CSDiag.cpp - Constraint Diagnostics ------------------------------===//
-//
-// This source file is part of the Swift.org open source project
-//
-// Copyright (c) 2014 - 2018 Apple Inc. and the Swift project authors
-// Licensed under Apache License v2.0 with Runtime Library Exception
-//
-// See https://swift.org/LICENSE.txt for license information
-// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
-//
-//===----------------------------------------------------------------------===//
-//
-// This file implements diagnostics for the type checker.
-//
-//===----------------------------------------------------------------------===//
-
-#include "CSDiag.h"
-#include "CSDiagnostics.h"
-#include "ConstraintSystem.h"
-#include "MiscDiagnostics.h"
-#include "TypeCheckAvailability.h"
-#include "TypoCorrection.h"
-#include "swift/AST/ASTWalker.h"
-#include "swift/AST/DiagnosticEngine.h"
-#include "swift/AST/GenericEnvironment.h"
-#include "swift/AST/Initializer.h"
-#include "swift/AST/ParameterList.h"
-#include "swift/AST/ProtocolConformance.h"
-#include "swift/AST/TypeMatcher.h"
-#include "swift/AST/TypeWalker.h"
-#include "swift/Basic/Defer.h"
-#include "swift/Basic/StringExtras.h"
-#include "llvm/ADT/DenseSet.h"
-#include "llvm/ADT/SetVector.h"
-#include "llvm/Support/Compiler.h"
-#include "llvm/Support/SaveAndRestore.h"
-
-using namespace swift;
-using namespace constraints;
-
-namespace swift {
-  Type replaceTypeParametersWithUnresolved(Type ty) {
-    if (!ty) return ty;
-    
-    if (!ty->hasTypeParameter() && !ty->hasArchetype()) return ty;
-    
-    auto &ctx = ty->getASTContext();
-    
-    return ty.transform([&](Type type) -> Type {
-      if (type->is<ArchetypeType>() ||
-          type->isTypeParameter())
-        return ctx.TheUnresolvedType;
-      return type;
-    });
-  }
-
-  Type replaceTypeVariablesWithUnresolved(Type ty) {
-    if (!ty) return ty;
-    
-    if (!ty->hasTypeVariable()) return ty;
-    
-    auto &ctx = ty->getASTContext();
-    
-    return ty.transform([&](Type type) -> Type {
-      if (type->isTypeVariableOrMember())
-        return ctx.TheUnresolvedType;
-      return type;
-    });
-  }
-};
-
-static bool isUnresolvedOrTypeVarType(Type ty) {
-  return ty->isTypeVariableOrMember() || ty->is<UnresolvedType>();
-}
-
-/// Flags that can be used to control name lookup.
-enum TCCFlags {
-  /// Allow the result of the subexpression to be an lvalue.  If this is not
-  /// specified, any lvalue will be forced to be loaded into an rvalue.
-  TCC_AllowLValue = 0x01,
-  
-  /// Re-type-check the given subexpression even if the expression has already
-  /// been checked already.  The client is asserting that infinite recursion is
-  /// not possible because it has relaxed a constraint on the system.
-  TCC_ForceRecheck = 0x02,
-    
-  /// tell typeCheckExpression that it is ok to produce an ambiguous result,
-  /// it can just fill in holes with UnresolvedType and we'll deal with it.
-  TCC_AllowUnresolvedTypeVariables = 0x04
-};
-
-using TCCOptions = OptionSet<TCCFlags>;
-
-inline TCCOptions operator|(TCCFlags flag1, TCCFlags flag2) {
-  return TCCOptions(flag1) | flag2;
-}
-
-
-namespace {
-/// If a constraint system fails to converge on a solution for a given
-/// expression, this class can produce a reasonable diagnostic for the failure
-/// by analyzing the remnants of the failed constraint system. (Specifically,
-/// left-over inactive, active and failed constraints.)
-/// This class does not tune its diagnostics for a specific expression kind,
-/// for that, you'll want to use an instance of the FailureDiagnosis class.
-class FailureDiagnosis :public ASTVisitor<FailureDiagnosis, /*exprresult*/bool>{
-  friend class ASTVisitor<FailureDiagnosis, /*exprresult*/bool>;
-  
-  Expr *expr = nullptr;
-  ConstraintSystem &CS;
-
-public:
-  FailureDiagnosis(Expr *expr, ConstraintSystem &cs) : expr(expr), CS(cs) {
-    assert(expr);
-  }
-
-  template<typename ...ArgTypes>
-  InFlightDiagnostic diagnose(ArgTypes &&...Args) {
-    return CS.getASTContext().Diags.diagnose(std::forward<ArgTypes>(Args)...);
-  }
-
-  /// Unless we've already done this, retypecheck the specified child of the
-  /// current expression on its own, without including any contextual
-  /// constraints or the parent expr nodes.  This is more likely to succeed than
-  /// type checking the original expression.
-  ///
-  /// This mention may only be used on immediate children of the current expr
-  /// node, because ClosureExpr parameters need to be treated specially.
-  ///
-  /// This can return a new expression (for e.g. when a UnresolvedDeclRef gets
-  /// resolved) and returns null when the subexpression fails to typecheck.
-  ///
-  Expr *typeCheckChildIndependently(
-      Expr *subExpr, Type convertType = Type(),
-      ContextualTypePurpose convertTypePurpose = CTP_Unused,
-      TCCOptions options = TCCOptions(),
-      ExprTypeCheckListener *listener = nullptr,
-      bool allowFreeTypeVariables = true);
-  Expr *typeCheckChildIndependently(Expr *subExpr, TCCOptions options,
-                                    bool allowFreeTypeVariables = true) {
-    return typeCheckChildIndependently(subExpr, Type(), CTP_Unused, options,
-                                       nullptr, allowFreeTypeVariables);
-  }
-
-  Type getTypeOfTypeCheckedChildIndependently(Expr *subExpr,
-                                            TCCOptions options = TCCOptions()) {
-    auto e = typeCheckChildIndependently(subExpr, options);
-    return e ? CS.getType(e) : Type();
-  }
-
-  /// Find a nearest declaration context which could be used
-  /// to type-check this sub-expression.
-  DeclContext *findDeclContext(Expr *subExpr) const;
-
-  void getPossibleTypesOfExpressionWithoutApplying(
-      Expr *&expr, DeclContext *dc, SmallPtrSetImpl<TypeBase *> &types,
-      FreeTypeVariableBinding allowFreeTypeVariables =
-          FreeTypeVariableBinding::Disallow,
-      ExprTypeCheckListener *listener = nullptr) {
-    TypeChecker::getPossibleTypesOfExpressionWithoutApplying(
-        expr, dc, types, allowFreeTypeVariables, listener);
-    CS.cacheExprTypes(expr);
-  }
-
-  Type getTypeOfExpressionWithoutApplying(
-      Expr *&expr, DeclContext *dc, ConcreteDeclRef &referencedDecl,
-      FreeTypeVariableBinding allowFreeTypeVariables =
-          FreeTypeVariableBinding::Disallow,
-      ExprTypeCheckListener *listener = nullptr) {
-    auto type =
-        TypeChecker::getTypeOfExpressionWithoutApplying(expr, dc,
-                                                        referencedDecl,
-                                                        allowFreeTypeVariables,
-                                                        listener);
-    CS.cacheExprTypes(expr);
-    return type;
-  }
-
-  /// Attempt to diagnose a specific failure from the info we've collected from
-  /// the failed constraint system.
-  bool diagnoseExprFailure();
-
-  /// Emit an ambiguity diagnostic about the specified expression.
-  void diagnoseAmbiguity(Expr *E);
-
-private:
-  /// Validate potential contextual type for type-checking one of the
-  /// sub-expressions, usually correct/valid types are the ones which
-  /// either don't have type variables or are not generic, because
-  /// generic types with left-over type variables or unresolved types
-  /// degrade quality of diagnostics if allowed to be used as contextual.
-  ///
-  /// \param contextualType The candidate contextual type.
-  /// \param CTP The contextual purpose attached to the given candidate.
-  ///
-  /// \returns Pair of validated type and it's purpose, potentially nullified
-  /// if it wasn't an appropriate type to be used.
-  std::pair<Type, ContextualTypePurpose>
-  validateContextualType(Type contextualType, ContextualTypePurpose CTP);
-
-  bool visitExpr(Expr *E);
-
-  bool visitRebindSelfInConstructorExpr(RebindSelfInConstructorExpr *E);
-};
-} // end anonymous namespace
-
-namespace {
-  class ExprTypeSaverAndEraser {
-    llvm::DenseMap<Expr*, Type> ExprTypes;
-    llvm::DenseMap<TypeLoc*, Type> TypeLocTypes;
-    llvm::DenseMap<Pattern*, Type> PatternTypes;
-    ExprTypeSaverAndEraser(const ExprTypeSaverAndEraser&) = delete;
-    void operator=(const ExprTypeSaverAndEraser&) = delete;
-  public:
-
-    ExprTypeSaverAndEraser(Expr *E) {
-      struct TypeSaver : public ASTWalker {
-        ExprTypeSaverAndEraser *TS;
-        TypeSaver(ExprTypeSaverAndEraser *TS) : TS(TS) {}
-        
-        std::pair<bool, Expr *> walkToExprPre(Expr *expr) override {
-          TS->ExprTypes[expr] = expr->getType();
-
-          SWIFT_DEFER {
-            assert((!expr->getType() || !expr->getType()->hasTypeVariable()
-                    // FIXME: We shouldn't allow these, either.
-                    || isa<LiteralExpr>(expr)) &&
-                   "Type variable didn't get erased!");
-          };
-
-          // Preserve module expr type data to prevent further lookups.
-          if (auto *declRef = dyn_cast<DeclRefExpr>(expr))
-            if (isa<ModuleDecl>(declRef->getDecl()))
-              return { false, expr };
-          
-          // Don't strip type info off OtherConstructorDeclRefExpr, because
-          // CSGen doesn't know how to reconstruct it.
-          if (isa<OtherConstructorDeclRefExpr>(expr))
-            return { false, expr };
-          
-          // If a literal has a Builtin.Int or Builtin.FP type on it already,
-          // then sema has already expanded out a call to
-          //   Init.init(<builtinliteral>)
-          // and we don't want it to make
-          //   Init.init(Init.init(<builtinliteral>))
-          // preserve the type info to prevent this from happening.
-          if (isa<LiteralExpr>(expr) && !isa<InterpolatedStringLiteralExpr>(expr) &&
-              !(expr->getType() && expr->getType()->hasError()))
-            return { false, expr };
-          
-          expr->setType(nullptr);
-
-          return { true, expr };
-        }
-        
-        // If we find a TypeLoc (e.g. in an as? expr), save and erase it.
-        bool walkToTypeLocPre(TypeLoc &TL) override {
-          if (TL.getTypeRepr() && TL.getType()) {
-            TS->TypeLocTypes[&TL] = TL.getType();
-            TL.setType(Type());
-          }
-          return true;
-        }
-        
-        std::pair<bool, Pattern*> walkToPatternPre(Pattern *P) override {
-          if (P->hasType()) {
-            TS->PatternTypes[P] = P->getType();
-            P->setType(Type());
-          }
-          return { true, P };
-        }
-        
-        // Don't walk into statements.  This handles the BraceStmt in
-        // non-single-expr closures, so we don't walk into their body.
-        std::pair<bool, Stmt *> walkToStmtPre(Stmt *S) override {
-          return { false, S };
-        }
-      };
-      
-      E->walk(TypeSaver(this));
-    }
-    
-    void restore() {
-      for (auto exprElt : ExprTypes)
-        exprElt.first->setType(exprElt.second);
-      
-      for (auto typelocElt : TypeLocTypes)
-        typelocElt.first->setType(typelocElt.second);
-      
-      for (auto patternElt : PatternTypes)
-        patternElt.first->setType(patternElt.second);
-      
-      // Done, don't do redundant work on destruction.
-      ExprTypes.clear();
-      TypeLocTypes.clear();
-      PatternTypes.clear();
-    }
-    
-    // On destruction, if a type got wiped out, reset it from null to its
-    // original type.  This is helpful because type checking a subexpression
-    // can lead to replacing the nodes in that subexpression.  However, the
-    // failed ConstraintSystem still has locators pointing to the old nodes,
-    // and if expr-specific diagnostics fail to turn up anything useful to say,
-    // we go digging through failed constraints, and expect their locators to
-    // still be meaningful.
-    ~ExprTypeSaverAndEraser() {
-      for (auto exprElt : ExprTypes)
-        if (!exprElt.first->getType())
-          exprElt.first->setType(exprElt.second);
-      
-      for (auto typelocElt : TypeLocTypes)
-        if (!typelocElt.first->getType())
-          typelocElt.first->setType(typelocElt.second);
-      
-      for (auto patternElt : PatternTypes)
-        if (!patternElt.first->hasType())
-          patternElt.first->setType(patternElt.second);
-    }
-  };
-} // end anonymous namespace
-
-/// Unless we've already done this, retypecheck the specified subexpression on
-/// its own, without including any contextual constraints or parent expr
-/// nodes.  This is more likely to succeed than type checking the original
-/// expression.
-///
-/// This can return a new expression (for e.g. when a UnresolvedDeclRef gets
-/// resolved) and returns null when the subexpression fails to typecheck.
-Expr *FailureDiagnosis::typeCheckChildIndependently(
-    Expr *subExpr, Type convertType, ContextualTypePurpose convertTypePurpose,
-    TCCOptions options, ExprTypeCheckListener *listener,
-    bool allowFreeTypeVariables) {
-
-  // If this sub-expression is currently being diagnosed, refuse to recheck the
-  // expression (which may lead to infinite recursion).  If the client is
-  // telling us that it knows what it is doing, then believe it.
-  if (!options.contains(TCC_ForceRecheck)) {
-    if (CS.isExprBeingDiagnosed(subExpr)) {
-      auto *savedExpr = CS.getExprBeingDiagnosed(subExpr);
-      if (subExpr == savedExpr)
-        return subExpr;
-
-      CS.cacheExprTypes(savedExpr);
-      return savedExpr;
-    }
-  }
-
-  // Mark current expression as about to be diagnosed.
-  CS.addExprForDiagnosis(subExpr, subExpr);
-
-  // Validate contextual type before trying to use it.
-  std::tie(convertType, convertTypePurpose) =
-      validateContextualType(convertType, convertTypePurpose);
-
-  // If we have no contextual type information and the subexpr is obviously a
-  // overload set, don't recursively simplify this.  The recursive solver will
-  // sometimes pick one based on arbitrary ranking behavior (e.g. like
-  // which is the most specialized) even then all the constraints are being
-  // fulfilled by UnresolvedType, which doesn't tell us anything.
-  if (convertTypePurpose == CTP_Unused &&
-      (isa<OverloadedDeclRefExpr>(subExpr->getValueProvidingExpr()))) {
-    return subExpr;
-  }
-
-  // Save any existing type data of the subexpr tree, and reset it to null in
-  // prep for re-type-checking the tree.  If things fail, we can revert the
-  // types back to their original state.
-  ExprTypeSaverAndEraser SavedTypeData(subExpr);
-  
-  // Store off the sub-expression, in case a new one is provided via the
-  // type check operation.
-  Expr *preCheckedExpr = subExpr;
-  
-  // Make sure that typechecker knows that this is an attempt
-  // to diagnose a problem.
-  TypeCheckExprOptions TCEOptions =
-      TypeCheckExprFlags::SubExpressionDiagnostics;
-
-  // Claim that the result is discarded to preserve the lvalue type of
-  // the expression.
-  if (options.contains(TCC_AllowLValue))
-    TCEOptions |= TypeCheckExprFlags::IsDiscarded;
-
-  // If there is no contextual type available, tell typeCheckExpression that it
-  // is ok to produce an ambiguous result, it can just fill in holes with
-  // UnresolvedType and we'll deal with it.
-  if ((!convertType || options.contains(TCC_AllowUnresolvedTypeVariables)) &&
-      allowFreeTypeVariables)
-    TCEOptions |= TypeCheckExprFlags::AllowUnresolvedTypeVariables;
-
-  // When we're type checking a single-expression closure, we need to reset the
-  // DeclContext to this closure for the recursive type checking.  Otherwise,
-  // if there is a closure in the subexpression, we can violate invariants.
-  auto *DC = findDeclContext(subExpr);
-  auto resultTy =
-      TypeChecker::typeCheckExpression(subExpr, DC,
-                                       TypeLoc::withoutLoc(convertType),
-                                       convertTypePurpose, TCEOptions,
-                                       listener, &CS);
-
-  CS.cacheExprTypes(subExpr);
-
-  // This is a terrible hack to get around the fact that typeCheckExpression()
-  // might change subExpr to point to a new OpenExistentialExpr. In that case,
-  // since the caller passed subExpr by value here, they would be left
-  // holding on to an expression containing open existential types but
-  // no OpenExistentialExpr, which breaks invariants enforced by the
-  // ASTChecker.
-  // Another reason why we need to do this is because diagnostics might pick
-  // constraint anchor for re-typechecking which would only have opaque value
-  // expression and not enclosing open existential, which is going to trip up
-  // sanitizer.
-  eraseOpenedExistentials(CS, subExpr);
-
-  // If recursive type checking failed, then an error was emitted.  Return
-  // null to indicate this to the caller.
-  if (!resultTy)
-    return nullptr;
-
-  // If we type checked the result but failed to get a usable output from it,
-  // just pretend as though nothing happened.
-  if (resultTy->is<ErrorType>()) {
-    subExpr = preCheckedExpr;
-    if (subExpr->getType())
-      CS.cacheType(subExpr);
-    SavedTypeData.restore();
-  }
-
-  if (preCheckedExpr != subExpr)
-    CS.addExprForDiagnosis(preCheckedExpr, subExpr);
-
-  return subExpr;
-}
-
-DeclContext *FailureDiagnosis::findDeclContext(Expr *subExpr) const {
-  if (auto *closure =
-          dyn_cast<ClosureExpr>(subExpr->getSemanticsProvidingExpr()))
-    return closure->getParent();
-
-  struct DCFinder : public ASTWalker {
-    DeclContext *DC, *CurrDC;
-    Expr *SubExpr;
-
-    DCFinder(DeclContext *DC, Expr *expr) : DC(DC), CurrDC(DC), SubExpr(expr) {}
-
-    std::pair<bool, Expr *> walkToExprPre(Expr *E) override {
-      if (E == SubExpr) {
-        DC = CurrDC;
-        return {false, nullptr};
-      }
-
-      if (auto *closure = dyn_cast<ClosureExpr>(E)) {
-        CurrDC = closure;
-        // If we have a ClosureExpr parent of the specified node, check to make
-        // sure none of its arguments are type variables.  If so, these type
-        // variables would be accessible to name lookup of the subexpression and
-        // may thus leak in.  Reset them to UnresolvedTypes for safe measures.
-        assert(llvm::all_of(*closure->getParameters(), [](const ParamDecl *PD) {
-          if (PD->hasInterfaceType()) {
-            auto paramTy = PD->getType();
-            return !(paramTy->hasTypeVariable() || paramTy->hasError());
-          }
-          return true;
-        }));
-      }
-
-      return {true, E};
-    }
-
-    Expr *walkToExprPost(Expr *E) override {
-      if (auto *closure = dyn_cast<ClosureExpr>(E)) {
-        assert(CurrDC == closure && "DeclContext imbalance");
-        CurrDC = closure->getParent();
-      }
-      return E;
-    }
-
-  } finder(CS.DC, subExpr);
-
-  expr->walk(finder);
-  return finder.DC;
-}
-
-bool FailureDiagnosis::
-visitRebindSelfInConstructorExpr(RebindSelfInConstructorExpr *E) {
-  // Don't walk the children for this node, it leads to multiple diagnostics
-  // because of how sema injects this node into the type checker.
-  return false;
-}
-
-bool FailureDiagnosis::visitExpr(Expr *E) {
-  // Check each of our immediate children to see if any of them are
-  // independently invalid.
-  bool errorInSubExpr = false;
-  
-  E->forEachImmediateChildExpr([&](Expr *Child) -> Expr* {
-    // If we already found an error, stop checking.
-    if (errorInSubExpr) return Child;
-
-    // Otherwise just type check the subexpression independently.  If that
-    // succeeds, then we stitch the result back into our expression.
-    if (typeCheckChildIndependently(Child, TCC_AllowLValue))
-      return Child;
-
-    // Otherwise, it failed, which emitted a diagnostic.  Keep track of this
-    // so that we don't emit multiple diagnostics.
-    errorInSubExpr = true;
-    return Child;
-  });
-  
-  // If any of the children were errors, we're done.
-  if (errorInSubExpr)
-    return true;
-  
-  // Otherwise, produce a more generic error.
-  return false;
-}
-
-
-bool FailureDiagnosis::diagnoseExprFailure() {
-  assert(expr);
-
-  // Our general approach is to do a depth first traversal of the broken
-  // expression tree, type checking as we go.  If we find a subtree that cannot
-  // be type checked on its own (even to an incomplete type) then that is where
-  // we focus our attention.  If we do find a type, we use it to check for
-  // contextual type mismatches.
-  return visit(expr);
-}
-
-
-/// Given a specific expression and the remnants of the failed constraint
-/// system, produce a specific diagnostic.
-///
-/// This is guaranteed to always emit an error message.
-///
-void ConstraintSystem::diagnoseFailureFor(SolutionApplicationTarget target) {
-  setPhase(ConstraintSystemPhase::Diagnostics);
-
-  SWIFT_DEFER { setPhase(ConstraintSystemPhase::Finalization); };
-
-  if (auto expr = target.getAsExpr()) {
-    // Look through RebindSelfInConstructorExpr to avoid weird Sema issues.
-    if (auto *RB = dyn_cast<RebindSelfInConstructorExpr>(expr))
-      expr = RB->getSubExpr();
-
-    FailureDiagnosis diagnosis(expr, *this);
-
-    // Now, attempt to diagnose the failure from the info we've collected.
-    if (diagnosis.diagnoseExprFailure())
-      return;
-
-    // If no one could find a problem with this expression or constraint system,
-    // then it must be well-formed... but is ambiguous.  Handle this by diagnostic
-    // various cases that come up.
-    diagnosis.diagnoseAmbiguity(expr);
-  } else {
-    // Emit a poor fallback message.
-    getASTContext().Diags.diagnose(
-         target.getAsFunction()->getLoc(), diag::failed_to_produce_diagnostic);
-  }
-}
-
-std::pair<Type, ContextualTypePurpose>
-FailureDiagnosis::validateContextualType(Type contextualType,
-                                         ContextualTypePurpose CTP) {
-  if (!contextualType)
-    return {contextualType, CTP};
-
-  // Since some of the contextual types might be tuples e.g. subscript argument
-  // is a tuple or paren wrapping a tuple, it's required to recursively check
-  // its elements to determine nullability of the contextual type, because it
-  // might contain archetypes.
-  std::function<bool(Type)> shouldNullifyType = [&](Type type) -> bool {
-    switch (type->getDesugaredType()->getKind()) {
-    case TypeKind::PrimaryArchetype:
-    case TypeKind::OpenedArchetype:
-    case TypeKind::NestedArchetype:
-    case TypeKind::Unresolved:
-      return true;
-
-    case TypeKind::BoundGenericEnum:
-    case TypeKind::BoundGenericClass:
-    case TypeKind::BoundGenericStruct:
-    case TypeKind::UnboundGeneric:
-    case TypeKind::GenericFunction:
-    case TypeKind::Metatype:
-      return type->hasUnresolvedType();
-
-    case TypeKind::Tuple: {
-      auto tupleType = type->getAs<TupleType>();
-      for (auto &element : tupleType->getElements()) {
-        if (shouldNullifyType(element.getType()))
-            return true;
-      }
-      break;
-    }
-
-    default:
-      return false;
-    }
-
-    return false;
-  };
-
-  bool shouldNullify = false;
-  if (auto objectType = contextualType->getWithoutSpecifierType()) {
-    // Note that simply checking for `objectType->hasUnresolvedType()` is not
-    // appropriate in this case standalone, because if it's in a function,
-    // for example, or inout type, we still want to preserve it's skeleton
-    /// because that helps to diagnose inout argument issues. Complete
-    // nullification is only appropriate for generic types with unresolved
-    // types or standalone archetypes because that's going to give
-    // sub-expression solver a chance to try and compute type as it sees fit
-    // and higher level code would have a chance to check it, which avoids
-    // diagnostic messages like `cannot convert (_) -> _ to (Int) -> Void`.
-    shouldNullify = shouldNullifyType(objectType);
-  }
-
-  // If the conversion type contains no info, drop it.
-  if (shouldNullify)
-    return {Type(), CTP_Unused};
-
-  // Remove all of the potentially leftover type variables or type parameters
-  // from the contextual type to be used by new solver.
-  contextualType = replaceTypeParametersWithUnresolved(contextualType);
-  contextualType = replaceTypeVariablesWithUnresolved(contextualType);
-
-  return {contextualType, CTP};
-}
-
-/// Emit an ambiguity diagnostic about the specified expression.
-void FailureDiagnosis::diagnoseAmbiguity(Expr *E) {
-  if (auto *assignment = dyn_cast<AssignExpr>(E)) {
-    if (isa<DiscardAssignmentExpr>(assignment->getDest())) {
-      auto *srcExpr = assignment->getSrc();
-      diagnoseAmbiguity(srcExpr);
-      return;
-    }
-  }
-
-  // Unresolved/Anonymous ClosureExprs are common enough that we should give
-  // them tailored diagnostics.
-  if (auto CE = dyn_cast<ClosureExpr>(E->getValueProvidingExpr())) {
-    diagnose(CE->getLoc(), diag::cannot_infer_closure_type)
-      .highlight(CE->getSourceRange());
-    return;
-  }
-
-  // Attempt to re-type-check the entire expression, allowing ambiguity, but
-  // ignoring a contextual type.
-  if (expr == E) {
-    auto exprType = getTypeOfTypeCheckedChildIndependently(expr);
-    // If it failed and diagnosed something, then we're done.
-    if (!exprType) return;
-
-    // If we were able to find something more specific than "unknown" (perhaps
-    // something like "[_:_]" for a dictionary literal), include it in the
-    // diagnostic.
-    if (!isUnresolvedOrTypeVarType(exprType)) {
-      diagnose(E->getLoc(), diag::specific_type_of_expression_is_ambiguous,
-               exprType)
-        .highlight(E->getSourceRange());
-      return;
-    }
-  }
-
-  // Before giving up completely let's try to see if there are any
-  // fixes recorded by constraint generator, which point to structural
-  // problems that might not result in solution even if fixed e.g.
-  // missing members involved in protocol composition in expression
-  // context which are interpreted as binary operator expressions instead.
-  {
-    bool diagnosed = false;
-    for (auto *fix : CS.getFixes())
-      diagnosed |= fix->diagnose();
-
-    if (diagnosed)
-      return;
-  }
-
-  // If there are no posted constraints or failures, then there was
-  // not enough contextual information available to infer a type for the
-  // expression.
-  diagnose(E->getLoc(), diag::type_of_expression_is_ambiguous)
-    .highlight(E->getSourceRange());
-}
