+++ test/Parse/matching_patterns.swift
@@ -334,4 +334,4 @@ case (_?)?: break // expected-warning {{case is already handled by previous patt
 let (responseObject: Int?) = op1
 // expected-error @-1 {{expected ',' separator}} {{25-25=,}}
 // expected-error @-2 {{expected pattern}}
-// expected-error @-3 {{expression type 'Int?' is ambiguous without more context}}
+// expected-error @-3 {{type of expression is ambiguous without more context}}
